Specify optional_T for arg type, instead of is_optional

+++ src/types_primitives_test.zig
@@ -79,7 +79,7 @@ const Primitives = struct {
     pub fn checkNullEmpty(_: Self, v: ?u32) bool {
         return (v == null);
     }
-    pub fn checkNullNotEmpty(_: Self, v: ?u32) bool {
+    pub fn checkNullNotEmpty(_: Self, v: ?u8) bool {
         return (v != null);
     }
 };
